requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
boto3==1.34.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.6.15
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
idna==3.10
jiter==0.10.0
openai==1.93.0
pydantic-core==2.14.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1

